arreglos

+++ backend/src/services/menu.service.js
@@ -287,6 +287,11 @@ export async function deleteMenuByIdService(id_menu) {
         });
         // Verificar si el menú existe
         if (!menuFound) return [null, { dataInfo: "id_menu", message: `El menú con ID ${id_menu} no existe.` }];
+
+        // Verificar si el menú está disponible
+        if (menuFound.disponibilidad) {
+            return [null, "No se puede eliminar un menú disponible. Desactívelo primero."];
+        }
         // Eliminar el menú
         const menuDeleted = await menuRepository.remove(menuFound);
